dj_domconnect/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
django==3.0.5
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
pyparsing==2.4.7
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
django-bootstrap-modal-forms
gunicorn==20.0.4
# psycopg2-binary==2.8.5
# djangorestframework-simplejwt
# djoser
